Feature/match attendant request

+++ domain/registration.go
@@ -2,8 +2,8 @@ package domain
 
 type Registration struct {
 	BaseModel
-	PlayerId int64  `json:"playerId"`
-	MatchId  int64  `json:"matchId"`
+	PlayerId uint   `json:"playerId"`
+	MatchId  uint   `json:"matchId"`
 	IsPaid   bool   `json:"isPaid"`
 	Comment  string `json:"comment"`
 }
